# openapi/context-brain-actions.public.openapi.yaml
openapi: "3.1.0"
info:
  title: "Context Brain Memory Lab — Public Agent API"
  description: |
    Public agent-facing REST API for Context Brain Memory Lab (OpenCB).

    This schema is a deliberately curated Actions subset of the full API
    surface, covering the routes safe for external agents, GPT Actions, and
    non-MCP integrations. Intentionally excluded (available on the live API
    and via MCP, see /openapi.json on a deployment): admin/tier-override,
    conflict escalations, inferred-edge approve/reject, reasoning
    traverse/explain/answer, context packs, batch save, similar retrieval,
    feedback, ask metrics, and keyword-audit routes. Exclusion here is a
    curation decision, not absence of the capability.

    **Auth:** Bearer token via `Authorization: Bearer <token>` header.
    Workspace scope is resolved server-side from the token; callers never
    inject raw workspace UUIDs unless the deployment grants explicit override.

    **Base URL:** configure per deployment (see `docs/GPT_ACTIONS.md`).

    **MCP alternative:** if your client supports MCP streamable-http transport,
    connect to `POST /mcp` with the same Bearer token — all 34 tools are
    available with identical semantics to this REST surface.
  version: "1.0.0"
  contact:
    name: "OpenCB project"
servers:
  - url: "https://your-opencb-host/api"
    description: "Self-hosted deployment (replace with your actual host)"
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]
...
    DecisionStatus:
      type: string
      enum: [active, superseded, reversed, draft]

    # ── Health ───────────────────────────────────────────────────────────────
...
    DecisionCreate:
      type: object
      required: [title, decision_reason]
      properties:
        title:
          type: string
        decision_reason:
          type: string
          description: "Why this decision was made."
        decision_context:
          type: string
          nullable: true
        why_this_matters:
          type: string
          nullable: true
        decision_status:
          $ref: "#/components/schemas/DecisionStatus"
          default: active
        reversible:
          type: boolean
          default: true
        confidence_level:
          type: string
          enum: [low, medium, high]
          default: medium
        decision_tags:
          type: array
          items:
            type: string
          nullable: true
        linked_hub_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
        supersedes_decision_id:
          type: string
          format: uuid
          nullable: true
        source_content_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
        alternatives_considered:
          type: array
          items:
            type: object
            properties:
              option:
                type: string
              reason_rejected:
                type: string
          nullable: true
        contradicting_evidence:
          type: string
          nullable: true
    DecisionSummary:
      type: object
      properties:
        decision_id:
          type: string
          format: uuid
        title:
          type: string
        decision_status:
          $ref: "#/components/schemas/DecisionStatus"
        confidence_level:
          type: string
        created_at:
          type: string
          format: date-time
    DecisionContentLink:
      allOf:
        - $ref: "#/components/schemas/DecisionSummary"
        - type: object
          properties:
            link_role:
              type: string
              enum: [canonical, source]
              description: "How the decision references the content: its canonical narrative representation, or declared supporting evidence."
            also_source:
              type: boolean
              description: "True when a canonical-role decision also lists the content among its sources."
    DecisionsByContentResponse:
      type: object
      properties:
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/DecisionContentLink"
        count:
          type: integer
        content_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
          nullable: true
    DecisionFull:
      allOf:
        - $ref: "#/components/schemas/DecisionSummary"
        - type: object
          properties:
            decision_reason:
              type: string
            decision_context:
              type: string
              nullable: true
            why_this_matters:
              type: string
              nullable: true
            reversible:
              type: boolean
            decision_tags:
              type: array
              items:
                type: string
            linked_hub_ids:
              type: array
              items:
                type: string
                format: uuid
            supersedes_decision_id:
              type: string
              format: uuid
              nullable: true
            alternatives_considered:
              type: array
              items:
                type: object
            contradicting_evidence:
              type: string
              nullable: true
    DecisionStatusUpdate:
      type: object
      required: [decision_status]
      properties:
        decision_status:
          $ref: "#/components/schemas/DecisionStatus"

    # ── Graph ────────────────────────────────────────────────────────────────
...
paths:
...
  /decisions/:
    get:
      operationId: list_decisions
      summary: "List decisions"
      tags: [decisions]
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DecisionStatus"
            nullable: true
        - name: hub_id
          in: query
          schema:
            type: string
            format: uuid
            nullable: true
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: "Decision list"
          content:
            application/json:
              schema:
                type: object
                properties:
                  decisions:
                    type: array
                    items:
                      $ref: "#/components/schemas/DecisionSummary"
                  count:
                    type: integer
    post:
      operationId: create_decision_memory
      summary: "Save a decision with reasoning"
      description: |
        Store an important architectural, product, or strategic decision
        with its full rationale. Future agents can retrieve *why* something
        was decided, not just what.
      tags: [decisions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionCreate"
      responses:
        "201":
          description: "Decision created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionSummary"
        "404":
          description: "Referenced hub or content not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /decisions/timeline:
    get:
      operationId: get_decision_timeline
      summary: "Decision history timeline"
      tags: [decisions]
      parameters:
        - name: hub_id
          in: query
          schema:
            type: string
            format: uuid
            nullable: true
        - name: tags
          in: query
          schema:
            type: string
            nullable: true
            description: "Comma-separated tag filter"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: "Timeline grouped by status"
          content:
            application/json:
              schema:
                type: object
  /decisions/{decision_id}:
    get:
      operationId: explain_decision
      summary: "Full decision record"
      description: "Retrieve full decision including reasoning, context, and why it matters."
      tags: [decisions]
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "Full decision"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionFull"
        "404":
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /decisions/by-content/{content_id}:
    get:
      operationId: list_decisions_for_content
      summary: "Which decisions reference this content item?"
      description: |
        The content→decision join (CF-002): given a content_id (e.g. a
        retrieval or ask evidence hit), returns the decision nodes that link
        it. link_role says how — canonical (the decision's own narrative
        representation) or source (caller-declared supporting evidence).
        count=0 means no decision references it: an unlinked item, not an
        error; unknown ids also return count=0 so the read never leaks
        whether content exists.
      tags: [decisions]
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: "Decisions linking the content item, newest first"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionsByContentResponse"
  /decisions/{decision_id}/lineage:
    get:
      operationId: get_decision_lineage
      summary: "Decision reasoning chain"
      description: "Get ancestors (what this replaced) and descendants (what replaced it)."
      tags: [decisions]
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "Lineage chain"
          content:
            application/json:
              schema:
                type: object
        "404":
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /decisions/{decision_id}/status:
    patch:
      operationId: update_decision_status
      summary: "Update decision lifecycle status"
      tags: [decisions]
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionStatusUpdate"
      responses:
        "200":
          description: "Updated decision"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionFull"
        "404":
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Graph ──────────────────────────────────────────────────────────────────
...